+++ src/main/java/roomescape/controller/AdminReservationTimeController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.CreateReservationTimeResponse;
-import roomescape.dto.ReservationTimeRequest;
+import roomescape.dto.response.CreateReservationTimeResponse;
+import roomescape.dto.request.ReservationTimeRequest;
 import roomescape.service.ReservationTimeService;
 
 @RestController
