--- src/main/java/roomescape/dto/CreateReservationTimeResponse.java
@@ -1,16 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
-
-public record CreateReservationTimeResponse(
-        Long id,
-        LocalTime startAt
-) {
-    public static CreateReservationTimeResponse from(ReservationTime reservationTime) {
-        return new CreateReservationTimeResponse(
-                reservationTime.getId(),
-                reservationTime.getStartAt()
-        );
-    }
-}
