--- src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,29 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalDate;
-import roomescape.domain.Reservation;
-
-public record ReservationResponse(
-        Long id,
-        String name,
-        LocalDate date,
-        CreateReservationTimeResponse time,
-        ThemeResponse theme
-) {
-    public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(
-                reservation.getId(),
-                reservation.getName(),
-                reservation.getDate(),
-               new CreateReservationTimeResponse(
-                       reservation.getTime().getId(),
-                       reservation.getTime().getStartAt()
-               ),
-                new ThemeResponse(
-                        reservation.getTheme().getId(),
-                        reservation.getTheme().getName(),
-                        reservation.getTheme().getDescription(),
-                        reservation.getTheme().getThumbnail())
-        );
-    }
-}
