--- src/main/java/roomescape/dto/ReservationTimeRequest.java
@@ -1,12 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
-
-public record ReservationTimeRequest(
-        LocalTime startAt
-) {
-    public ReservationTime toReservationTime() {
-        return ReservationTime.createWithoutId(startAt);
-    }
-}
