--- src/main/java/roomescape/dto/ThemeResponse.java
@@ -1,14 +1,0 @@
-package roomescape.dto;
-
-import roomescape.domain.Theme;
-
-public record ThemeResponse(
-        Long id,
-        String name,
-        String description,
-        String thumbnail
-) {
-    public static ThemeResponse from(Theme theme) {
-        return new ThemeResponse(theme.getId(), theme.getName(), theme.getDescription(), theme.getThumbnail());
-    }
-}
