+++ src/main/java/roomescape/dto/request/ReservationRequest.java
@@ -1,15 +1,13 @@
 package roomescape.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 
 public record ReservationRequest(
-        @NotBlank String name,
-        @NotNull LocalDate date,
+        String name,
+        LocalDate date,
         long timeId,
         long themeId
