+++ src/main/java/roomescape/dto/request/ReservationTimeRequest.java
@@ -1,11 +1,10 @@
 package roomescape.dto.request;
 
-import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
 
 public record ReservationTimeRequest(
-        @NotNull LocalTime startAt
+        LocalTime startAt
 ) {
     public ReservationTime toReservationTime() {
         return ReservationTime.createWithoutId(startAt);
