+++ src/main/java/roomescape/dto/request/ThemeRequest.java
@@ -1,13 +1,11 @@
 package roomescape.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
-import org.hibernate.validator.constraints.URL;
 import roomescape.domain.Theme;
 
 public record ThemeRequest(
-        @NotBlank String name,
-        @NotBlank String description,
-        @URL String thumbnail
+        String name,
+        String description,
+        String thumbnail
 ) {
     public Theme toTheme() {
         return Theme.createWithoutId(
